# Tensor squash kernel optimizations