Optimization changes of Op2dTensorSquash kernel.  I see that performance improvement around 64% to 65% in new kernel compared with current implementation for below config.  Here are performance improvement numbers.

Original code:
root@hpe-sjc2-43:/testx/MIOpen/hip_build# ./bin/MIOpenDriver rnnfp16 -n 224 -W 224 -H 1000 -l 8 -b 1 -m lstm -p 0 -r 0 -k 32 -F 4 -t 1 -w 1 -V 0 -i 100
MIOpenDriver rnnfp16 -n 224 -W 224 -H 1000 -l 8 -b 1 -m lstm -p 0 -r 0 -k 32 -F 4 -t 1 -w 1 -V 0 -i 100
length of data sequence == 1 is short than time sequence == 32, padding the rest of data sequence with 224
length of data sequence == 1 is short than time sequence == 32, padding the rest of data sequence with 224
GPU Kernel Time Backward Weights RNN Elapsed: 33.102783 ms
Wall-clock Time Backward Weights RNN Elapsed: 34.182896 ms

Optimization code:
root@hpe-sjc2-43:/testx/MIOpen/hip_build# ./bin/MIOpenDriver rnnfp16 -n 224 -W 224 -H 1000 -l 8 -b 1 -m lstm -p 0 -r 0 -k 32 -F 4 -t 1 -w 1 -V 0 -i 100
MIOpenDriver rnnfp16 -n 224 -W 224 -H 1000 -l 8 -b 1 -m lstm -p 0 -r 0 -k 32 -F 4 -t 1 -w 1 -V 0 -i 100
length of data sequence == 1 is short than time sequence == 32, padding the rest of data sequence with 224
length of data sequence == 1 is short than time sequence == 32, padding the rest of data sequence with 224
GPU Kernel Time Backward Weights RNN Elapsed: 11.397395 ms
Wall-clock Time Backward Weights RNN Elapsed: 12.437284 ms