@junliume 2022-04-06
> > > Hi @muralinr please check the CI issues with related tests. We need to cherry-pick this change to 5.1.x too, so first step is to push it through CI and merge it into develop. Thanks!
> > 
> > 
> > Sure Jun. Let me take a look at these failures.
> 
> I see these tests pass on MI200 target. Which target do you see these failures?
> 
> root@hpe-sjc2-43:/testx/MIOpen/build# ./bin/test_rnn_vanilla_dropout Empty batch sequence. Filling uniformly with batch size: 17 root@hpe-sjc2-43:/testx/MIOpen/build# ./bin/test_rnn_vanilla_dropout --all Empty batch sequence. Filling uniformly with batch size: 17 Empty batch sequence. Filling uniformly with batch size: 17 Empty batch sequence. Filling uniformly with batch size: 17 Empty batch sequence. Filling uniformly with batch size: 17 Empty batch sequence. Filling uniformly with batch size: 17 Empty batch sequence. Filling uniformly with batch size: 17 Empty batch sequence. Filling uniformly with batch size: 17 Empty batch sequence. Filling uniformly with batch size: 17 root@hpe-sjc2-43:/testx/MIOpen/build# ./bin/test_lstm_dropout --all Empty batch sequence. Filling uniformly with batch size: 17 Empty batch sequence. Filling uniformly with batch size: 17 Empty batch sequence. Filling uniformly with batch size: 17 Empty batch sequence. Filling uniformly with batch size: 17

NODE_NAME = ixt-rack-100
I think it is a gfx900 or gfx906