@muralinr 2022-04-06
> 

Hi Jun, I don't have gfx900 or gfx906 machine. Can I get one node on these machines?